refactor: separate use nft dashboard to separate file
